Split test_update_project.py into focused modules
Address CodeScene Low Cohesion warning (5 responsibilities > 4 threshold) on tests/setup-cmd/test_update_project.py by splitting its 8 test classes across 6 focused modules and extracting shared helpers into _update_project_helpers.py:

- test_update_project_validation.py    -- TestValidation
- test_update_project_missing_files.py  -- TestMissingFileCopy + TestEmptyDiffSkip
- test_update_project_sha_reading.py    -- TestPerFileShaReading
- test_update_project_first_sync.py     -- TestFirstSync
- test_update_project_routine.py        -- TestRoutineUpdate + TestAbortOnFailure
- test_update_project_idempotence.py    -- TestIdempotenceAndCleanup

Also extract a helper inside TestEmptyDiffSkip to remove structural duplication CodeScene flagged after the split. All 58 setup-cmd tests pass; each new file scores 10.0 on Code Health.

Co-authored by Claude Code
